+++ Week9/src/pages/Cart/hooks/useClearCart.ts
@@ -1,0 +1,17 @@
+import { clearCart } from '@/store/cartSlice';
+import { useAppDispatch } from '@/store/hooks';
+import { closeModal, openModal } from '@/store/modalSlice';
+
+const dispatch = useAppDispatch();
+
+export const useClearCart = () => {
+  dispatch(
+    openModal({
+      message: '정말 삭제하시겠습니까?',
+      onConfirm: () => {
+        dispatch(clearCart());
+        dispatch(closeModal());
+      },
+    }),
+  );
+};
